+++ benchmarks/benchjs/multiop.js
@@ -85,7 +85,7 @@ mongoose.connect('mongodb://localhost/mongoose-bench', function(err) {
         body: 'texttt'
       }]
     };
-    for (var i = 0; i < 5; i++) {
+    for (let i = 0; i < 5; i++) {
       blogData.comments.push(commentData);
     }
     const UserSchema = new Schema({
@@ -115,7 +115,7 @@ mongoose.connect('mongodb://localhost/mongoose-bench', function(err) {
 
     // insert all of the data here
     let count = 4000;
-    for (i = 0; i < 1000; i++) {
+    for (let i = 0; i < 1000; i++) {
       data.age = Math.floor(Math.random() * 50);
       User.create(data, function(err, u) {
         if (err) {